+++ src/packages/core/auth/auth-provider.extension.ts
