+++ src/packages/core/auth/index.ts
@@ -5,5 +5,6 @@ export * from './auth.context.token.js';
 export * from './modals/index.js';
 export * from './models/openApiConfiguration.js';
 export * from './components/index.js';
+export * from './auth-provider.extension.js';
 
 export type * from './types.js';
