+++ src/packages/core/extension-registry/extensions/index.ts
@@ -1,5 +1,4 @@
 export type * from './app-entry-point.extension.js';
-export type * from './auth-provider.extension.js';
 export type * from './backoffice-entry-point.extension.js';
 export type * from './entity-action.extension.js';
 export type * from './entity-bulk-action.extension.js';
